backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleak>=1.1.0",
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "uvicorn[standard]>=0.35.0",
]
